ref: refactored the page.tsx for better readability
