+++ src/components/layout/main.tsx
@@ -1,7 +1,85 @@
+import { Button } from "@/components/ui/button";
+import {
+  Cloud,
+  Code,
+  Cpu,
+  Database,
+  Globe,
+  Server,
+  Shield,
+  LucideIcon,
+} from "lucide-react";
 export default function Main() {
     return (
       <section className="relative bg-[#020817] py-24 overflow-hidden mt-20">
         <div className="container mx-auto px-4">
+        <main className="container mx-auto px-4 py-20 text-center relative">
+        <p className="text-lg text-gray-300 mb-4">
+          Secure. Scalable. Always Reliable
+        </p>
+        <h1 className="text-4xl md:text-5xl lg:text-4xl font-bold mb-6 bg-gradient-to-r from-blue-500 to-white text-transparent bg-clip-text">
+          Empower your growth with lightning-fast, secure,
+          <br />
+          and customization cloud technology
+        </h1>
+        <p className="text-gray-400 max-w-2xl mx-auto mb-12">
+          Take control of your cloud. Deploy fast, scale instantly, and pay as
+          you go.
+          <br />
+          Enjoy top performance, strong security, and 24/7 reliability
+        </p>
+
+        {/* Cloud Services Diagram - Now using relative units */}
+        <div
+          className="relative w-full max-w-[90vw] aspect-square mx-auto mb-12"
+          style={{ maxWidth: "600px" }}
+        >
+          {/* Center Icon */}
+          <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2">
+            <div className="w-32 h-32  rounded-full flex items-center justify-center">
+              <Cloud className="w-16 h-16 text-blue-500" />
+            </div>
+          </div>
+
+          {/* Orbital Icons - Using percentage-based positioning */}
+          <div className="absolute inset-0">
+            {/* Top */}
+            <div className="absolute top-[5%] left-1/2 transform -translate-x-1/2">
+              <IconCircle Icon={Shield} />
+            </div>
+            {/* Top Right */}
+            <div className="absolute top-[25%] right-[5%]">
+              <IconCircle Icon={Server} />
+            </div>
+            {/* Bottom Right */}
+            <div className="absolute bottom-[25%] right-[5%]">
+              <IconCircle Icon={Database} />
+            </div>
+            {/* Bottom */}
+            <div className="absolute bottom-[5%] left-1/2 transform -translate-x-1/2">
+              <IconCircle Icon={Code} />
+            </div>
+            {/* Bottom Left */}
+            <div className="absolute bottom-[25%] left-[5%]">
+              <IconCircle Icon={Globe} />
+            </div>
+            {/* Top Left */}
+            <div className="absolute top-[25%] left-[5%]">
+              <IconCircle Icon={Cpu} />
+            </div>
+          </div>
+
+          {/* Decorative Circles - Using percentage-based sizing */}
+          <div className="absolute inset-0">
+            <div className="absolute top-1/2 left-1/2 w-[80%] h-[80%] transform -translate-x-1/2 -translate-y-1/2 border border-blue-500/20 rounded-full" />
+            <div className="absolute top-1/2 left-1/2 w-[60%] h-[60%] transform -translate-x-1/2 -translate-y-1/2 border border-blue-500/30 rounded-full" />
+          </div>
+        </div>
+
+        <Button className="bg-blue-600 hover:bg-blue-700 text-white px-8 py-6 text-lg rounded-full">
+          Explore All Solution →
+        </Button>
+      </main>
           <div className="grid grid-cols-1 lg:grid-cols-2 gap-12 items-center">
             {/* Text Content */}
             <div className="space-y-6">
@@ -46,8 +124,9 @@ export default function Main() {
         </div>
   
         {/* Background Glow Effect */}
-        <div className="absolute inset-0 bg-[radial-gradient(ellipse_at_center,_var(--tw-gradient-stops))] from-blue-900/20 via-gray-900 to-gray-900 opacity-50" />
+        
       </section>
+    
     )
   }
   
@@ -60,4 +139,10 @@ export default function Main() {
     )
   }
   
-  
+  function IconCircle({ Icon }: { Icon: LucideIcon }) {
+    return (
+      <div className="w-[10%] aspect-square min-w-[40px] bg-blue-600 rounded-full flex items-center justify-center shadow-lg shadow-blue-500/50 hover:shadow-blue-500/75 transition-shadow duration-300">
+        <Icon className="w-1/2 h-1/2 text-white" />
+      </div>
+    );
+  }
